+++ src/Impact/Embodied/Engine.php
@@ -57,7 +57,7 @@ class Engine extends CommonGLPI
     }
 
     /**
-     * Get an instance of the engine to calculate imapcts for the given itemtype
+     * Get an instance of the engine to calculate impacts for the given itemtype
      *
      * Returns null if no engine found
      *
@@ -73,6 +73,7 @@ class Engine extends CommonGLPI
         }
 
         $embodied_impact_namespace = Config::getEmbodiedImpactEngine();
+        /** @var class-string<AbstractEmbodiedImpact> $embodied_impact_class */
         $embodied_impact_class = $embodied_impact_namespace . '\\' . $itemtype;
         $must_implement = AbstractEmbodiedImpact::class;
         if (!class_exists($embodied_impact_class) || !is_subclass_of($embodied_impact_class, $must_implement)) {